prototype

+++ ASE.UI/Controllers/StaffController.cs
@@ -5,21 +5,20 @@ using System.Web;
 using System.Web.Mvc;
 using ASE.Entities;
 using ASE.BL;
+using ASE.UI;
 using System.IO;
 
 namespace CareerPanorama
 {
+	[Authorize]
     public class StaffController : ASE.BaseController<object>
 	{
         // GET: Staff
         public ActionResult Index()
         {
-			Guid id = Guid.Parse("97c11d5a-9ec0-4b91-9dbe-dc8520d307c7");
-			var staffmanager = new StaffManager();
-			List<StaffEntity> staff = staffmanager.GetStaffByUserIDEntities(id);
-			if (staff.Count == 0)
-				return View("Error");
-					
+			Guid id = UIHelper.GetUserIDByEmail(User.Identity.Name);
+			if (!IsAuthorized())
+				return View("Error");			
 			var stallmanager = new StallManager();
 			List<StallEntity> list = stallmanager.GetStallByUserIDEntities(id);
 			return View("StallSelection", list);
@@ -27,15 +26,24 @@ namespace CareerPanorama
 
 		public ActionResult Stall(Guid StallId)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			return View("StallOrders", StallId);
 		}
 
 		public ActionResult RemoveOrder(Guid ID)
-		{	
+		{
+			if (!IsAuthorized())
+				return View("Error");
 			var _manager = new OrderManager();
-			var stallid = _manager.GetOrderByIDObject(ID).Fields.StallID;
-			_manager.DeleteOrderByID(ID);
-			return RedirectToAction("Stall", new { StallID = stallid });
+			var stall = _manager.GetOrderByIDObject(ID);
+			if (stall != null)
+			{
+				var stallid = _manager.GetOrderByIDObject(ID).Fields.StallID;
+				_manager.DeleteOrderByID(ID);
+				return RedirectToAction("Stall", new { StallID = stallid });
+			}
+			return RedirectToAction("Index");
 		}
 
 		public PartialViewResult RefreshOrders(Guid StallID)
@@ -45,12 +53,16 @@ namespace CareerPanorama
 
 		public ActionResult CreateStall()
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			return View();
 		}
 
 		[HttpPost]
 		public ActionResult CreateStall(string Name, string Description, HttpPostedFileBase Logo)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			Stall obj = new Stall();
 			obj.Fields.Name = Name;
 			obj.Fields.Description = Description;
@@ -63,6 +75,8 @@ namespace CareerPanorama
 
 		public ActionResult Dishes(Guid StallId)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			var _manager = new DishManager();
 			var list = _manager.GetDishByStallIDEntities(StallId);
 			ViewData["StallId"] = StallId;
@@ -71,6 +85,8 @@ namespace CareerPanorama
 
 		public ActionResult CreateDish(Guid StallId)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			ViewData["StallId"] = StallId;
 			return View("CreateDish");
 		}
@@ -80,6 +96,8 @@ namespace CareerPanorama
 		[HttpPost]
 		public ActionResult SubmitDish(Dish dish)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			//using (var binaryReader = new BinaryReader(Logo.InputStream))
 			//	obj.Fields.Logo = binaryReader.ReadBytes(Logo.ContentLength);
 			dish.Save();
@@ -88,6 +106,8 @@ namespace CareerPanorama
 
 		public ActionResult RemoveDish(Guid DishId)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			var _manager = new DishManager();
 			_manager.DeleteDishByID(DishId);
 			return RedirectToAction("Index");
@@ -95,6 +115,8 @@ namespace CareerPanorama
 
 		public ActionResult Users()
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			var _manager = new StaffManager();
 			var staff = _manager.GetAllStaffEntities();
 			return View("Users", staff);
@@ -102,8 +124,11 @@ namespace CareerPanorama
 
 		public ActionResult AddStaffByEmail(string Email)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			var _manager = new UserManager();
-			if (_manager.GetUserIdByEmail(Email) != null)
+			var userid = _manager.GetUserIdByEmail(Email);
+			if (userid != null && userid != Guid.Empty)
 			{
 				var obj = new Staff(Email);
 				obj.Save();
@@ -113,10 +138,26 @@ namespace CareerPanorama
 
 		public ActionResult RemoveStaff(Guid StaffID)
 		{
+			if (!IsAuthorized())
+				return View("Error");
 			var _manager = new StaffManager();
 			_manager.DeleteStaffByID(StaffID);
 			return RedirectToAction("Users");
 		}
+
+		#region Helper Functions
+
+		public bool IsAuthorized()
+		{
+			Guid id = UIHelper.GetUserIDByEmail(User.Identity.Name);
+			var staffmanager = new StaffManager();
+			List<StaffEntity> staff = staffmanager.GetStaffByUserIDEntities(id);
+			if (staff.Count == 0)
+				return false;
+			return true;
+		}
+
+		#endregion
 
     }
 }
